Functions: Upgrade node, packages; fix typescript errors
To get the functions building and deployable on Firebase, a number of
package upgrades were needed, which also triggered upgrading Typescript,
which triggered needing to update some code.

Also had to update tsconfig, as the new aws/client-sdk library wasn't
compliling due to not being able to import json, and due to unused
locals.