Convert auto-importer to writing all answers for run_key
This mirrors what the full-copy export-answers script does.

Whenever an answer changes, we get the run_key for that answer, which is
equivalent to the intersection of the platform_user_id and resource_link_id,
and save a sync doc with that run_key.

When it is time to write (when the timer monitorSyncDocCount fires), we
find all answers with that run_key and append them to a single parquet
file.

Previously, we were blowing away the complete parquet file for a run_key
every time an answer changed, because we were replacing it with one that
just had the single answer.

An advantage of this is that we now will only fire off the writing of a
single parquet file for all the changes a user makes during the four
minutes.

One issue is that writing may take longer. In my testing, with a few
answers including one with a large model serialization, it took 28
seconds to write the parquet file, and a further 11 seconds to send it
to S3.