Replace run_key with LTI identifiers for logged-in users
When we have a logged-in user, we expect to also have a platform_id, a
resource_link_id and a platform_user_id. These three identifiers together
uniquely identify a learner-assignment pair.

This updates the sync_doc and answer-finding code to support these.

This also changes the storage location of the Parquet files in S3,
putting them in nested folders that allow each learner-assignment to
create a unique file. For anonymous users, this goes in a new
"anonymous" folder.