Update export-answers for LTI and new folder structure
This echoes the changes in auto-importer.ts to default to keying the
answers from the LTI info if possible, and to upload the parquet files
to the new folder structure.

All the new logic is imported from the shared ts library.

This also changes the script so that the bucket and source names are
brought in as arguments when calling the script.