Redirect to researcher-reports, set in env vars
This redirects to the researcher-reports url after the queries have all
been started.

The template and samconfigs have been changed so that, by default, when
a dev is running it redirects to `https://localhost:8080`, and when
deploying to staging and production it redirects to
`https://researcher-reports.concord.org/branch/master/` and
`https://researcher-reports.concord.org` respectively.