Include portal param in redirect url
This way the query-creator app passes the appropriate portal to the
researcher-report, which then uses it for auth.

If this manner of getting the portal url from the learner API url ends
up being flaky, we could instead just include it as another url parameter
in the query-creator app, like the reportServiceSource, but it seems like
that would be redundant.