use column abstraction
This way the columns for the header row and the columns for the main select should always line up.
This also makes the use of AS more consistent.
The tests had to be updated for this use of AS.

This should also pave the way for breaking the grouping sub query into two parts and moving
it to the WITH clause