Generate model url and add to SQL results
This pieces together a url to the Portal Report to open up a student's
version of a model. The Portal Report will authenticate the user before
showing the model.

This relies on the offering_id being in the learner JSON (there is a
PR in the portal for this) and in the learners table (see new SQL for
generating the learners table).