feat: Add project admin/researcher access [PT-188646004]
NOTE: a migration required using mix ecto.migrate to add new columns.

This adds the ability for project admins and researchers to access the
reports, and adds new columns to the user table to denote if the user is
a project admin or researcher.

When the report form is loaded a list of allowed project ids is read
from the database based on the user's access level.  For admins this
list is set to the atom :all and for users without any access it is set
to :none.

The report filter options queries have been updated to filter the
results based on the allowed project ids if the user is not an admin.

TODO LIST:

1. Update the report queries to filter based on the allowed project ids
2. Update the new report form to first check option counts before
   telling the user to type in the first 3 characters.