fix: Move /reports to /old-reports [PT-188779638]
This is the first step in a three step process to remove the old reports and replace the /reports url with the new reports.

Once this is live the existing query-creator Firebase function can be updated to redirect to /old-reports.  Once that is live then a follow on PR will rename /new-reports to /reports.

Then, the portals can then be updated to remove the report links from the admin and add the report link in the sidebar to point to /reports.

Finally, a follow on PR will remove all the /old-reports code that is no longer needed.