fix: Exclude CC users learner data query [REPORT-29]
Previously the query was using a "NOT LIKE '%@concord.org'" filter which both caused a full table scan of the users table (which is very large in production).  This caused the query to timeout.

The change is to instead query the database separately if the exclude CC users setting is set to get a list of all teacher ids in schools with "Concord Consortium" in their name and then join against the portal teachers table and exclude those teachers from the result.  This was the original behavior of the exclude CC users checkbox in the old reports.